pyproject.toml
[tool.poetry]
name = "multimodalinterp"
version = "0.1.0"
description = ""
authors = ["Francesc0rtu <francescortu@live.it>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
torch = "2.4.1"
pyarrow="18.1.0" 
datasets = "2.20.0"
packaging = "^24.1"
nltk = "3.8.1"
matplotlib = "^3.9.0"
inflect = "^7.3.0"
easyroutine = "^0.3"
python-dotenv = "^1.0.1"
scipy = "^1.14.0"
seaborn = "^0.13.2"
streamlit = "^1.36.0"
numpy = "^1.26.4"
plotly = "^5.23.0"
line-profiler = "^4.1.3"
pandas = "^2.2.2"
jaxtyping = "0.2.0"
dadapy = {git = "https://github.com/sissa-data-science/DADApy"}
ipykernel = "^6.29.5"
gputil = "^1.4.0"
accelerate = "^0.34.2"
tiktoken = "^0.8.0"
timm = "^1.0.11"
peft = "^0.14.0"
xformers = "0.0.28"
sentencepiece = "^0.2.0"
statsmodels = "^0.14.4"
torchvision = "^0.19"
tokenizers = "^0.21.0"
flax = "^0.8.5"
opencv-python = "^4.10.0.84"
loguru = "^0.7.3"
rich = "^13.9.4"
memory-profiler = "^0.61.0"
transformers = {git = "https://github.com/francescortu/transformers", rev = "NarrowGate_3.0"}




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"